fix(profile): update file location
