+++ Client/src/modules/profile/view/component/EmailPasswordPhoneSetting.jsx
@@ -1,5 +1,3 @@
-import { updateUserPassword } from '@/api/profile/updateUserPassword';
-import { updateUserProfile } from '@/api/profile/updateUserProfile';
 import { Button } from '@/components/ui/button';
 import { Card, CardDescription, CardHeader } from '@/components/ui/card';
 import {
@@ -13,6 +11,8 @@ import {
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { useToast } from '@/components/ui/use-toast';
+import { updateUserPassword } from '@/modules/profile/data/source/updateUserPassword';
+import { updateUserProfile } from '@/modules/profile/data/source/updateUserProfile';
 import PropTypes from 'prop-types';
 import { useState } from 'react';
